Refractored skyblock to not use inner classes
