+++ src/main/java/io/github/neopixel/HypixelAPI.java
@@ -3,8 +3,8 @@ package io.github.neopixel;
 import io.github.neopixel.http.RequestController;
 import io.github.neopixel.wrapper.guild.HypixelGuild;
 import io.github.neopixel.wrapper.player.HypixelPlayer;
-import io.github.neopixel.wrapper.skyblock.SkyblockElection;
-import io.github.neopixel.wrapper.skyblock.SkyblockNews;
+import io.github.neopixel.wrapper.skyblock.election.SkyblockElection;
+import io.github.neopixel.wrapper.skyblock.news.SkyblockNews;
 import io.github.neopixel.wrapper.skyblock.player.SkyblockProfile;
 import java.io.File;
 import java.util.UUID;
